--- code/.eslintignore
@@ -1,6 +1,0 @@
-node_modules/
-coverage/
-.idea/
-.next/
-.vscode/
-build/
