--- code/next.config.js
@@ -1,8 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  experimental: {
-    ppr: false
-  }
-};
-
-module.exports = nextConfig;
