+++ code/package.json
@@ -9,26 +9,32 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/node": "20.10.1",
-    "@types/react": "18.2.39",
-    "@types/react-dom": "18.2.17",
-    "autoprefixer": "10.4.16",
-    "eslint-config-next": "canary",
-    "next": "canary",
-    "postcss": "8.4.31",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "3.3.5",
-    "typescript": "5.3.2"
+    "autoprefixer": "^10.4.20",
+    "next": "15.0.0-canary.120",
+    "postcss": "^8.4.41",
+    "react": "19.0.0-rc-4c2e457c7c-20240522",
+    "react-dom": "19.0.0-rc-4c2e457c7c-20240522",
+    "tailwindcss": "^3.4.10"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "canary",
-    "@vercel/style-guide": "^5.1.0",
-    "eslint": "8.54.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.29.0",
-    "eslint-plugin-prettier": "^5.0.1",
-    "prettier": "^3.1.0",
-    "prettier-plugin-tailwindcss": "^0.5.7"
+    "@eslint/compat": "^1.1.1",
+    "@next/eslint-plugin-next": "15.0.0-canary.120",
+    "@types/node": "^22.4.0",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vercel/style-guide": "^6.0.0",
+    "babel-plugin-react-compiler": "0.0.0-experimental-592953e-20240517",
+    "eslint": "^9.9.0",
+    "eslint-config-next": "15.0.0-canary.120",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react-compiler": "0.0.0-experimental-c8b3f72-20240517",
+    "globals": "^15.9.0",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.6",
+    "typescript": "^5.5.4",
+    "typescript-eslint": "^8.1.0"
   }
 }
